+++ src/main/java/Ping/Main.java
@@ -1,5 +1,4 @@
-package Ping;
-
+package ping;
 import java.io.IOException;
 
 import javafx.application.Application;
@@ -7,11 +6,15 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import ping.exceptions.PingException;
+import ping.gui.MainWindow;
+
 
 /**
  * A GUI for Duke using FXML.
  */
 public class Main extends Application {
+    private static final String BOT_NAME = "Ping ChatBot";
 
     private Ping ping = new Ping();
 
@@ -24,6 +27,7 @@ public class Main extends Application {
             FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
+            stage.setTitle(BOT_NAME);
             stage.setScene(scene);
             fxmlLoader.<MainWindow>getController().setPing(ping);
             stage.show();
