--- src/main/java/Ping/MainWindow.java
@@ -1,63 +1,0 @@
-package Ping;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.image.Image;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
-/**
- * Controller for MainWindow. Provides the layout for the other controls.
- */
-public class MainWindow extends AnchorPane {
-    @FXML
-    private ScrollPane scrollPane;
-    @FXML
-    private VBox dialogContainer;
-    @FXML
-    private TextField userInput;
-    @FXML
-    private Button sendButton;
-
-    private Ping ping;
-
-    private Image userImage = new Image(this.getClass().getResourceAsStream("/images/LineFace.png"));
-    private Image pingImage = new Image(this.getClass().getResourceAsStream("/images/Ping.png"));
-
-    @FXML
-    public void initialize() {
-        scrollPane.vvalueProperty().bind(dialogContainer.heightProperty());
-    }
-
-    public void setPing(Ping d) {
-        ping = d;
-    }
-
-    /**
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    @FXML
-    private void handleUserInput() throws PingException {
-        String input = userInput.getText();
-        String response = ping.getResponse(input);
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(input, userImage),
-                DialogBox.getDukeDialog(response, pingImage)
-        );
-        if (input.equals("bye")) {
-            // close the window after 1 second
-            new java.util.Timer().schedule(
-                    new java.util.TimerTask() {
-                        @Override
-                        public void run() {
-                            System.exit(0);
-                        }
-                    },
-                    1000
-            );
-        }
-        userInput.clear();
-    }
-}
