+++ src/main/java/Ping/Parser.java
@@ -1,19 +1,27 @@
-package Ping;
+package ping;
 
-import java.util.List;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
-import Command.AddCommand;
-import Command.BlahCommand;
-import Command.Command;
-import Command.DeleteCommand;
-import Command.ExitCommand;
-import Command.FindCommand;
-import Command.HiCommand;
-import Command.ListCommand;
-import Command.MarkCommand;
-import Command.UnmarkCommand;
+import ping.command.AddCommand;
+import ping.command.BlahCommand;
+import ping.command.Command;
+import ping.command.DeleteCommand;
+import ping.command.ExitCommand;
+import ping.command.FindCommand;
+import ping.command.HiCommand;
+import ping.command.ListCommand;
+import ping.command.MarkCommand;
+import ping.command.UnmarkCommand;
+import ping.exceptions.PingException;
+import ping.job.Deadline;
+import ping.job.Event;
+import ping.job.Todo;
+import ping.timematch.DateTimeCheck;
+
+
+
 
 /**
  * This class is used to parse the command input by the user
