+++ src/main/java/Ping/Storage.java
@@ -1,4 +1,5 @@
-package Ping;
+package ping;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
@@ -6,6 +7,10 @@ import java.io.IOException;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
 import java.util.ArrayList;
+
+import ping.exceptions.PingException;
+import ping.job.Task;
+
