+++ src/main/java/Ping/job/Todo.java
@@ -1,12 +1,19 @@
 //// Solution below adapted by week2 iP Level-3 Partial solution
-package Ping;
+package ping.job;
 
 /**
  * This class is used to create a todo task
  */
 public class Todo extends Task {
+    protected String description;
+
     public Todo(String description) {
         super(description);
+    }
+
+
+    public String getDescription() {
+        return description;
     }
 
     @Override
