+++ src/main/resources/view/MainWindow.fxml
@@ -2,18 +2,30 @@
 
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.SplitPane?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Ping.MainWindow">
+<AnchorPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ping.gui.MainWindow">
     <children>
-        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0" AnchorPane.bottomAnchor="1.0" />
-        <Button fx:id="sendButton" layoutX="324.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="41.0" prefWidth="76.0" text="Send" />
-        <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="400.0" vvalue="1.0">
-            <content>
-                <VBox fx:id="dialogContainer" prefHeight="552.0" prefWidth="388.0" />
-            </content>
-        </ScrollPane>
+      <BorderPane prefHeight="600.0" prefWidth="800.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
+         <center>
+              <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="400.0" prefWidth="800.0" style="-fx-background-color: #F6F6F7;" vvalue="1.0" BorderPane.alignment="CENTER">
+               <content>
+                      <VBox fx:id="dialogContainer" prefHeight="544.0" prefWidth="797.0" />
+               </content>
+              </ScrollPane>
+         </center>
+         <bottom>
+            <SplitPane dividerPositions="0.8" nodeOrientation="LEFT_TO_RIGHT" prefHeight="50.0" prefWidth="800.0" BorderPane.alignment="CENTER">
+               <items>
+                    <TextField fx:id="userInput" alignment="TOP_LEFT" onAction="#handleUserInput" prefHeight="50.0" prefWidth="674.0" />
+                    <Button fx:id="sendButton" alignment="CENTER" contentDisplay="CENTER" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="50.0" prefWidth="117.0" text="Send" />
+               </items>
+            </SplitPane>
+         </bottom>
+      </BorderPane>
     </children>
 </AnchorPane>
