+++ src/test/java/Ping/TodoTest.java
@@ -1,13 +1,15 @@
-package Ping;
+package ping;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import ping.job.Todo;
 public class TodoTest {
     @Test
     public void todoCommand_ShouldHaveCorrectFormat() {
         Todo todo = new Todo("shower");
-        assertEquals("shower", todo.description);
+        assertEquals("shower", todo.getDescription());
     }
 
     @Test
