--- text-ui-test/EXPECTED.TXT
@@ -1,23 +1,0 @@
-Hello! I'm Ping
-What can I do for you?
-No ability added yet
-added: blah
-added: read book
-added: return book
-1.[ ] blah
-2.[ ] read book
-3.[ ] return book
-Nice! I've marked this task as done:
-[X] read book
-Got it. I've added this task:
-[D][ ] return book (by: Sunday)
-Now you have 4 tasks in the list.
-Got it. I've added this task:
-[E][ ] project meeting (from: Mon 2pm to: 4pm)
-Now you have 5 tasks in the list.
-1.[ ] blah
-2.[X] read book
-3.[ ] return book
-4.[D][ ] return book (by: Sunday)
-5.[E][ ] project meeting (from: Mon 2pm to: 4pm)
-Bye. Hope to see you again soon!
