--- text-ui-test/input.txt
@@ -1,10 +1,0 @@
-list
-blah
-read book
-return book
-list
-mark 2
-deadline return book /by Sunday
-event project meeting /from Mon 2pm /to 4pm
-list
-bye
